[IMP] merge purchase_request_payment_type
