--- purchase_request_hide_create_po_button/__manifest__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    "name": "Purchase Request Hide Create PO Button",
-    "version": "16.0.1.0.0",
-    "category": "Purchase Management",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "depends": ["purchase_request"],
-    "data": [
-        "views/purchase_request_views.xml"
-    ],
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
