--- purchase_request_hide_create_po_button/models/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import purchase_request
