--- purchase_request_hide_create_po_button/views/purchase_request_views.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-  <odoo>
-
-    <!-- View purchase.request form -->
-    <record id="view_purchase_request_form" model="ir.ui.view">
-        <field name="name">view.purchase.request.form (Hide create PO button)</field>
-        <field name="model">purchase.request</field>
-        <field name="inherit_id" ref="purchase_request.view_purchase_request_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//form" position="inside">
-                <field name="hide_create_po_button" invisible="1" />
-            </xpath>
-            <xpath expr="//button[@name='%(purchase_request.action_purchase_request_line_make_purchase_order)d']" position="attributes">
-                <attribute name="states"/>
-                <attribute name="attrs">{'invisible':[('hide_create_po_button', '=', True)]}</attribute>
-            </xpath>
-        </field>
-    </record>
-
-  </odoo>
