+++ purchase_request_kmitl/i18n/th.po
@@ -2184,3 +2184,42 @@ msgstr ""
 #: model_terms:ir.ui.view,arch_db:purchase_request_kmitl.view_purchase_request_form
 msgid "รายละเอียดคำขอซื้อขอจ้าง"
 msgstr ""
+
+#. module: purchase_request_kmitl
+#: model:ir.model.fields.selection,name:purchase_request_kmitl.selection__purchase_request__payment_type__direct
+msgid "Direct paid"
+msgstr "จ่ายตรง"
+
+#. module: purchase_request_kmitl
+#: model:ir.model.fields.selection,name:purchase_request_kmitl.selection__purchase_request__payment_type__loan
+msgid "Loan"
+msgstr "เงินยืม"
+
+#. module: purchase_request_kmitl
+#: model:ir.model.fields,field_description:purchase_request_kmitl.field_purchase_request__payment_type
+msgid "Payment Type"
+msgstr "ประเภทการจ่ายเงิน"
+
+#. module: purchase_request_kmitl
+#: model:ir.model.fields.selection,name:purchase_request_kmitl.selection__purchase_request__payment_type__prepaid
+msgid "Prepaid"
+msgstr "สำรองจ่าย"
+
+#. module: purchase_request_kmitl
+#: model:ir.model,name:purchase_request_kmitl.model_purchase_request
+msgid "Purchase Request"
+msgstr "คำขอซื้อขอจ้าง"
+
+#. module: purchase_request_kmitl
+#: model_terms:ir.ui.view,arch_db:purchase_request_kmitl.view_purchase_request_form
+msgid ""
+"When payment type is Prepaid, please select the advance\n"
+"                            payer's name in the Vendor field."
+msgstr "กรณีเลือกประเภทการจ่ายเงินเป็นสำรองจ่ายให้เลือกชื่อผู้สำรองจ่าย"
+
+#. module: purchase_request_kmitl
+#: model_terms:ir.ui.view,arch_db:purchase_request_kmitl.view_purchase_request_form
+msgid ""
+"When payment type is Loan, please select the loan's\n"
+"                            name in the Vendor field."
+msgstr "กรณีเลือกประเภทการจ่ายเงินเป็นเงินยืมให้เลือกชื่อยืมเงิน"
