+++ purchase_request_kmitl/wizards/purchase_request_line_make_purchase_order.py
@@ -17,6 +17,7 @@ class PurchaseRequestLineMakePurchaseOrder(models.TransientModel):
             "operating_unit_id": self.item_ids.request_id.operating_unit_id.id,
             "account_fiscal_year_id": self.item_ids.request_id.account_fiscal_year_id.id,
             "requesting_operating_unit_id": self.item_ids.request_id.operating_unit_id.id,
+            "payment_type": self.item_ids.request_id.payment_type,
         })
         return vals
