--- purchase_request_payment_type/__manifest__.py
@@ -1,16 +1,0 @@
-{
-    "name": "Purchase Request Payment Type",
-    "version": "16.0.1.0.0",
-    "category": "Purchase Management",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "depends": [
-        "purchase_request_kmitl",
-        "purchase_order_payment_type",
-        "purchase_order_link_purchase_request",
-    ],
-    "data": ["data/purchase_exception.xml", "views/purchase_request_views.xml"],
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
