--- purchase_request_payment_type/i18n/th.po
@@ -1,55 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* purchase_request_payment_type
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-10-27 15:17+0000\n"
-"PO-Revision-Date: 2025-10-27 15:17+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: purchase_request_payment_type
-#: model:ir.model.fields.selection,name:purchase_request_payment_type.selection__purchase_request__payment_type__direct
-msgid "Direct paid"
-msgstr "จ่ายตรง"
-
-#. module: purchase_request_payment_type
-#: model:ir.model.fields.selection,name:purchase_request_payment_type.selection__purchase_request__payment_type__loan
-msgid "Loan"
-msgstr "เงินยืม"
-
-#. module: purchase_request_payment_type
-#: model:ir.model.fields,field_description:purchase_request_payment_type.field_purchase_request__payment_type
-msgid "Payment Type"
-msgstr "ประเภทการจ่ายเงิน"
-
-#. module: purchase_request_payment_type
-#: model:ir.model.fields.selection,name:purchase_request_payment_type.selection__purchase_request__payment_type__prepaid
-msgid "Prepaid"
-msgstr "สำรองจ่าย"
-
-#. module: purchase_request_payment_type
-#: model:ir.model,name:purchase_request_payment_type.model_purchase_request
-msgid "Purchase Request"
-msgstr "คำขอซื้อขอจ้าง"
-
-#. module: purchase_request_payment_type
-#: model_terms:ir.ui.view,arch_db:purchase_request_payment_type.view_purchase_request_form
-msgid ""
-"When payment type is Prepaid, please select the advance\n"
-"                            payer's name in the Vendor field."
-msgstr "กรณีเลือกประเภทการจ่ายเงินเป็นสำรองจ่ายให้เลือกชื่อผู้สำรองจ่าย"
-
-#. module: purchase_request_payment_type
-#: model_terms:ir.ui.view,arch_db:purchase_request_payment_type.view_purchase_request_form
-msgid ""
-"When payment type is Loan, please select the loan's\n"
-"                            name in the Vendor field."
-msgstr "กรณีเลือกประเภทการจ่ายเงินเป็นเงินยืมให้เลือกชื่อยืมเงิน"
