--- purchase_request_payment_type/models/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import purchase_request
