--- purchase_request_payment_type/views/purchase_request_views.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <!-- View purchase.request form -->
-    <record id="view_purchase_request_form" model="ir.ui.view">
-        <field name="name">view.purchase.request.form (Payment Type)</field>
-        <field name="model">purchase.request</field>
-        <field
-            name="inherit_id"
-            ref="purchase_request_kmitl.view_purchase_request_form"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='procurement_method_id']" position="after">
-                <field
-                    name="payment_type"
-                    required="1"
-                    attrs="{'readonly': [('is_editable','=', False)]}"
-                    widget="radio"
-                    options="{'horizontal': true}"
-                />
-            </xpath>
-            <xpath expr="//field[@name='line_ids']" position="before">
-                <group name="warning_prepaid">
-                    <group>
-                        <div
-                            colspan="2"
-                            class="alert alert-warning show"
-                            role="alert"
-                            attrs="{'invisible': [('payment_type', '!=', 'prepaid')]}"
-                        >
-                            When payment type is Prepaid, please select the advance
-                            payer's name in the Vendor field.
-                        </div>
-                    </group>
-                </group>
-
-                <group name="warning_loan">
-                    <group>
-                        <div
-                            colspan="2"
-                            class="alert alert-warning show"
-                            role="alert"
-                            attrs="{'invisible': [('payment_type', '!=', 'loan')]}"
-                        >
-                            When payment type is Loan, please select the loan's
-                            name in the Vendor field.
-                        </div>
-                    </group>
-                </group>
-            </xpath>
-        </field>
-    </record>
-
-    <!-- View purchase.request tree -->
-    <record id="view_purchase_request_tree" model="ir.ui.view">
-        <field name="name">view.purchase.request.tree (Payment Type)</field>
-        <field name="model">purchase.request</field>
-        <field
-            name="inherit_id"
-            ref="purchase_request_kmitl.view_purchase_request_tree"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='procurement_method_id']" position="after">
-                <field name="payment_type" optional="show" />
-            </xpath>
-        </field>
-    </record>
-
-    <record id="view_purchase_request_search" model="ir.ui.view">
-        <field name="name">purchase.request.list.select (Payment Type)</field>
-        <field name="model">purchase.request</field>
-        <field name="priority">30</field>
-        <field name="inherit_id" ref="purchase_request.view_purchase_request_search" />
-        <field name="arch" type="xml">
-            <field name="product_id" position="before">
-                <field name="payment_type" />
-            </field>
-        </field>
-    </record>
-
-    <!-- UX override: move payment_type into the purchase_request_group (defined in
-         purchase_request_kmitl.view_purchase_request_form_ux). This view inherits from
-         the UX view so it is applied right after the UX restructure. -->
-    <record id="view_purchase_request_form_ux" model="ir.ui.view">
-        <field name="name">view.purchase.request.form.ux (Payment Type)</field>
-        <field name="model">purchase.request</field>
-        <field
-            name="inherit_id"
-            ref="purchase_request_kmitl.view_purchase_request_form"
-        />
-        <field name="arch" type="xml">
-            <!-- Hide payment_type from its orphaned position -->
-            <xpath expr="//field[@name='payment_type']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <!-- Re-add payment_type into the custom group, before operating_unit_id -->
-            <xpath
-                expr="//group[@name='purchase_request_group']//field[@name='operating_unit_id']"
-                position="before"
-            >
-                <field
-                    name="payment_type"
-                    required="1"
-                    attrs="{'readonly': [('is_editable','=', False)]}"
-                    widget="radio"
-                    options="{'horizontal': true}"
-                />
-            </xpath>
-        </field>
-    </record>
-</odoo>
