--- purchase_request_payment_type/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import purchase_request_line_make_purchase_order
